move gallery view emitter to shared package
