+++ apps/web/pages/[username]/[collectionId]/index.tsx
@@ -2,7 +2,6 @@ import { GetServerSideProps } from 'next';
 import { route } from 'nextjs-routes';
 import { graphql, useLazyLoadQuery } from 'react-relay';
 
-import GalleryViewEmitter from '~/components/internal/GalleryViewEmitter';
 import { CollectionNavbar } from '~/contexts/globalLayout/GlobalNavbar/CollectionNavbar/CollectionNavbar';
 import { StandardSidebar } from '~/contexts/globalLayout/GlobalSidebar/StandardSidebar';
 import { CollectionIdQuery } from '~/generated/CollectionIdQuery.graphql';
@@ -10,6 +9,7 @@ import { MetaTagProps } from '~/pages/_app';
 import GalleryRedirect from '~/scenes/_Router/GalleryRedirect';
 import GalleryRoute from '~/scenes/_Router/GalleryRoute';
 import CollectionGalleryPage from '~/scenes/CollectionGalleryPage/CollectionGalleryPage';
+import GalleryViewEmitter from '~/shared/components/GalleryViewEmitter';
 import { openGraphMetaTags } from '~/utils/openGraphMetaTags';
 
 type CollectionGalleryProps = MetaTagProps & {
