+++ apps/web/src/scenes/NftDetailPage/NftDetailPage.tsx
@@ -11,13 +11,13 @@ import FullPageLoader from '~/components/core/Loader/FullPageLoader';
 import transitions, {
   ANIMATED_COMPONENT_TRANSLATION_PIXELS_LARGE,
 } from '~/components/core/transitions';
-import GalleryViewEmitter from '~/components/internal/GalleryViewEmitter';
 import ErrorBoundary from '~/contexts/boundary/ErrorBoundary';
 import { NftDetailPageFragment$key } from '~/generated/NftDetailPageFragment.graphql';
 import { NftDetailPageQuery } from '~/generated/NftDetailPageQuery.graphql';
 import { NftDetailPageQueryFragment$key } from '~/generated/NftDetailPageQueryFragment.graphql';
 import useKeyDown from '~/hooks/useKeyDown';
 import NotFound from '~/scenes/NotFound/NotFound';
+import GalleryViewEmitter from '~/shared/components/GalleryViewEmitter';
 import { useTrack } from '~/shared/contexts/AnalyticsContext';
 import { removeNullValues } from '~/shared/relay/removeNullValues';
